moved constants to ./config
