+++ deploy/deploy-fee-taker.js
@@ -1,7 +1,7 @@
 const { deployAndGetContractWithCreate3, deployAndGetContract } = require('@1inch/solidity-utils');
 const hre = require('hardhat');
 const { ethers, getChainId } = hre;
-const constants = require('./constants');
+const constants = require('../config/constants');
 
 module.exports = async ({ deployments, getNamedAccounts }) => {
     const networkName = hre.network.name;
